RELATED: ONE-4105 Provide universal convertor to the legacy DrillContext
