+++ package.json
@@ -122,7 +122,7 @@
     "@gooddata/goodstrap": "^68.6.0",
     "@gooddata/js-utils": "^3.10.5",
     "@gooddata/numberjs": "^3.2.4",
-    "@gooddata/typings": "^2.24.0",
+    "@gooddata/typings": "^2.24.2",
     "@types/mapbox-gl": "^1.6.1",
     "ag-grid-community": "20.0.0",
     "ag-grid-react": "20.0.0",
