+++ src/components/visualizations/headline/HeadlineTransformation.tsx
@@ -1,4 +1,4 @@
-// (C) 2007-2019 GoodData Corporation
+// (C) 2007-2020 GoodData Corporation
 import * as React from "react";
 import { WrappedComponentProps, injectIntl } from "react-intl";
 import noop = require("lodash/noop");
@@ -10,7 +10,8 @@ import {
     buildDrillEventData,
     getHeadlineData,
 } from "./utils/HeadlineTransformationUtils";
-import { convertHeadlineDrillIntersectionToLegacy, fireDrillEvent } from "../utils/drilldownEventing";
+import { fireDrillEvent } from "../utils/drilldownEventing";
+import { convertDrillContextToLegacy } from "../utils/drilldownEventingLegacy";
 import { IHeadlineTransformationProps } from "./types";
 
 /**
@@ -79,18 +80,9 @@ class HeadlineTransformation extends React.Component<IHeadlineTransformationProp
             executionResponse,
         );
         const { executionContext, drillContext } = drillEventDataExtended;
-        const { type, element, value } = drillContext;
         const drillEventDataOld: IDrillEvent = {
             executionContext,
-            drillContext: {
-                type,
-                element,
-                value,
-                intersection: convertHeadlineDrillIntersectionToLegacy(
-                    drillContext.intersection,
-                    executionRequest.afm,
-                ),
-            },
+            drillContext: convertDrillContextToLegacy(drillContext, executionContext),
         };
         fireDrillEvent(onFiredDrillEvent, drillEventDataOld, target);
     }
